requirements/local.txt
-r base.txt
Markdown==2.5.1
django-autofixture==0.9.1
django-extensions==1.4.0
django-filter==0.8
djangorestframework==2.4.3

